Refactor tests, split them in different files
